chore: splitting up the instructions
